Fix error when 2week given end date outside range
Problem: If two week was given a `viewEndDate` in its `data()` method
that was outside of the "days" array it constructs, it would then error
when rendering because `viewIndex` (which is set according to
`viewEndDate`'s position in the "days" array) would be -1.

Fix: After constructing the "days" array, check that `viewEndDate` is in
range. If it's not, default to the last day.